+++ packages/easy_onvif/lib/src/model/common/metadata_configuration_options.g.dart
@@ -13,9 +13,12 @@ MetadataConfigurationOptions _$MetadataConfigurationOptionsFromJson(
   maxContentFilterSize: OnvifUtil.optionalInt(
     json['@MaxContentFilterSize'] as String?,
   ),
-  ptzStatusFilterOptions: PtzStatusFilterOptions.fromJson(
-    json['PTZStatusFilterOptions'] as Map<String, dynamic>,
-  ),
+  ptzStatusFilterOptions:
+      json['PTZStatusFilterOptions'] == null
+          ? null
+          : PtzStatusFilterOptions.fromJson(
+            json['PTZStatusFilterOptions'] as Map<String, dynamic>,
+          ),
   extension:
       json['Extension'] == null
           ? null