+++ packages/easy_onvif/lib/src/model/media1/get_metadata_configuration_response.g.dart
@@ -9,9 +9,12 @@ part of 'get_metadata_configuration_response.dart';
 GetMetadataConfigurationResponse _$GetMetadataConfigurationResponseFromJson(
   Map<String, dynamic> json,
 ) => GetMetadataConfigurationResponse(
-  configuration: MetadataConfiguration.fromJson(
-    json['Configuration'] as Map<String, dynamic>,
-  ),
+  configuration:
+      json['Configuration'] == null
+          ? null
+          : MetadataConfiguration.fromJson(
+            json['Configuration'] as Map<String, dynamic>,
+          ),
 );
 
 Map<String, dynamic> _$GetMetadataConfigurationResponseToJson(