+++ packages/easy_onvif/tool/grind.dart
@@ -1,6 +1,6 @@
 import 'package:publish_tools/publish_tools.dart';
 
-main(args) async {
+Future<void> main(List<String> args) async {
   PublishTools.addAllTasks();
 
   grind(args);
@@ -8,7 +8,7 @@ main(args) async {
 
 @DefaultTask('Just keeping it real')
 @Depends('pt-commit', 'pt-publish')
-done() {
+void done() {
   log('commit to faithoflifedev/easy_onvif complete');
   log('publish to pub.dev/packages complete.');
   // log('commit to faithoflifedev/homebrew-easy_onvif complete');